% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_comment.R
\name{get_issue_comments}
\alias{get_issue_comments}
\title{FUNCTION_TITLE}
\usage{
get_issue_comments(repo, number, endpoint = "api.github.com",
  PAT = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{repo}{PARAM_DESCRIPTION}

\item{number}{PARAM_DESCRIPTION}

\item{endpoint}{PARAM_DESCRIPTION, Default: 'api.github.com'}

\item{PAT}{PARAM_DESCRIPTION, Default: Sys.getenv("GITHUB_PAT")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
