% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issues.R
\name{get_issues}
\alias{get_issues}
\title{FUNCTION_TITLE}
\usage{
get_issues(repo, last = 100, endpoint = "api.github.com",
  PAT = Sys.getenv("GITHUB_PAT"))
}
\arguments{
\item{repo}{PARAM_DESCRIPTION}

\item{last}{PARAM_DESCRIPTION}

\item{endpoint}{PARAM_DESCRIPTION, Default: 'api.github.com'}

\item{PAT}{PARAM_DESCRIPTION, Default: Sys.getenv("GITHUB_PAT")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
