% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issue_table.R
\name{issue_tibble}
\alias{issue_tibble}
\title{FUNCTION_TITLE}
\usage{
issue_tibble(obj)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
