% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_table.R
\name{md_table}
\alias{md_table}
\title{FUNCTION_TITLE}
\usage{
md_table(obj)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
